rename sidechain -> sidechain_interaction
